Fixing progress types

+++ BadgeUpClient/Types/Progress.cs
@@ -3,7 +3,7 @@
 	public class Progress
 	{
 		public bool IsComplete { get; set; }
-		public int PercentComplete { get; set; }
+		public float PercentComplete { get; set; }
 		public Group ProgressTree { get; set; }
 		public string AchievementId { get; set; }
 		public string EarnedAchievementId { get; set; }
